refactor: rename and move ClassDetail components
