+++ src/components/classDetail/carousel/Carousel.tsx
@@ -5,7 +5,7 @@ import {
   CarouselItem,
   CarouselNext,
   CarouselPrevious,
-} from "../ui/carousel";
+} from "@/components/ui/carousel";
 
 export interface ClassImage {
   id: number;
